搜索功能可显示出搜索结果了;数据库中user中的point改为float类型

+++ src/com/tutoring/dao/UserDAOImpl.java
@@ -40,7 +40,7 @@ public class UserDAOImpl extends HibernateDaoSupport implements UserDAO{
 		person.setFirstName(firstname);
 		person.setLastName(lastname);
 		person.setType(type);
-		person.setPoint(0);
+		person.setPoint((float) 0);
 		if(type=='1')
 		{
 			Student stu = new Student();
